--- lib/debug.js
@@ -1,3 +1,0 @@
-const debug = require('./../debug');
-
-module.exports = (path) => debug(`lib/${path}`);
