--- lib/dest-folder.js
@@ -1,2 +1,0 @@
-module.exports = (date) => {
-};
