--- lib/get-set-md5.js
@@ -1,9 +1,0 @@
-const md5File = require('md5-file');
-
-const debug = require('./debug')('lib/get-set-md5');
-
-module.exports = (filePath) => {
-    if (fs.existsSync(`${filePath}.md`)) {
-    } else {
-    }
-};
