--- lib/prepend-zero.js
@@ -1,7 +1,0 @@
-module.exports = (value, len = 2) => {
-    let returnValue = value.toString();
-    while (returnValue.length < len) {
-        returnValue = `0${returnValue}`;
-    }
-    return returnValue;
-};
