+++ src/counter/error.ts
@@ -1,8 +1,6 @@
-class CounterError extends Error {
+export default class CounterError extends Error {
   constructor(message) {
     super(message);
     this.name = `Danosaure.PhotoTimeline.${this.constructor.name}`;
   }
 }
-
-module.exports = CounterError;
