+++ src/folder-walker.ts
@@ -2,7 +2,7 @@ import 'babel-regenerator-runtime';
 import fs from 'fs';
 import path from 'path';
 
-function* walker(item) {
+function* walker(item:string):IterableIterator<string> {
   const stats = fs.statSync(item);
   if (stats.isDirectory()) {
     // eslint-disable-next-line no-restricted-syntax
