--- src/process-file/debug.js
@@ -1,3 +1,0 @@
-import debug from '../debug';
-
-export default (info) => debug([__dirname, info]);
