+++ src/process-file/debug.ts
@@ -1,0 +1,3 @@
+import debug from '../debug';
+
+export default (info:string) => debug([__dirname, info]);
